+++ src/util.ts
@@ -1,21 +1,27 @@
 import * as vscode from 'vscode';
 
 // create status bar button
-export function statButton(name: string, command: string, align = vscode.StatusBarAlignment.Left, priority = 10) {
+export function resgisterStatButton(
+    context: vscode.ExtensionContext,
+    name: string,
+    command: string,
+    align = vscode.StatusBarAlignment.Left,
+    priority = 10) {
+
     let stat = vscode.window.createStatusBarItem(align, priority);
     stat.text = name;
     stat.command = command;
     stat.show();
-    return stat;
+    context.subscriptions.push(stat);
 }
 
 // create a terminal and send command
-export function simplTerm(name: string, cmd: string) {
-    return () => {
+export function registerSimplTerm(context: vscode.ExtensionContext, command: string, name: string, cmd: string) {
+    context.subscriptions.push(vscode.commands.registerCommand(command, () => {
         let t = vscode.window.createTerminal(name);
         t.sendText(cmd);
         t.show();
-    };
+    }));
 }
 
 // is a haskell file